# GymPlatform/docs/openapi.json
{"openapi":"3.1.0","info":{"title":"GymPlatform API","description":"API REST multi-tenant para administración de gimnasios. Autenticación JWT. Usa POST /api/auth/login para obtener el token. Al crear un gimnasio o un usuario staff sin contraseña, la clave inicial es 12345678.","contact":{"name":"GymPlatform","email":"admin@gymplatform.com"},"version":"1.0.0"},"servers":[{"url":"http://localhost:51820","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Gimnasio","description":"Usuarios, paquetes, actividades, reservaciones y rutinas"},{"name":"Autenticación","description":"Login y registro de miembros"},{"name":"Plataforma","description":"Gestión de clientes (gimnasios) — solo PLATFORM_OWNER"},{"name":"Público","description":"Endpoints sin autenticación"}],"paths":{"/api/users/{id}":{"put":{"tags":["Gimnasio"],"summary":"Actualizar usuario del gimnasio","description":"Actualiza datos de un usuario del gimnasio. Password opcional: si se omite, no cambia.","operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/users/me/profile":{"put":{"tags":["Gimnasio"],"operationId":"updateMyProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemberProfileUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/routine-requests/{id}/status":{"put":{"tags":["Gimnasio"],"operationId":"updateRoutineRequestStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoutineRequestStatusUpdate"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoutineRequestResponse"}}}}}}},"/api/platform/organizations/{id}":{"get":{"tags":["Plataforma"],"operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrganizationResponse"}}}}}},"put":{"tags":["Plataforma"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrganizationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrganizationResponse"}}}}}}},"/api/packages/{id}":{"get":{"tags":["Gimnasio"],"operationId":"getPackage","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MembershipPackageResponse"}}}}}},"put":{"tags":["Gimnasio"],"operationId":"updatePackage","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MembershipPackageRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MembershipPackageResponse"}}}}}}},"/api/users":{"get":{"tags":["Gimnasio"],"operationId":"getUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"post":{"tags":["Gimnasio"],"summary":"Crear usuario del gimnasio","description":"Crea usuarios del gimnasio actual con uno o más roles (Admin, Instructor, Miembro). Cada usuario pertenece solo a su organización. Password opcional (default 12345678).","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/routines":{"post":{"tags":["Gimnasio"],"operationId":"createRoutine","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoutineRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoutineResponse"}}}}}}},"/api/routines/assign-template":{"post":{"tags":["Gimnasio"],"operationId":"assignTemplate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignTemplateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoutineResponse"}}}}}}}},"/api/routine-templates":{"get":{"tags":["Gimnasio"],"operationId":"getTemplates","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoutineTemplateResponse"}}}}}}},"post":{"tags":["Gimnasio"],"operationId":"createTemplate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoutineTemplateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoutineTemplateResponse"}}}}}}},"/api/routine-requests":{"get":{"tags":["Gimnasio"],"operationId":"getRoutineRequests","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoutineRequestResponse"}}}}}}},"post":{"tags":["Gimnasio"],"operationId":"createRoutineRequest","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoutineRequestCreate"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoutineRequestResponse"}}}}}}},"/api/reservations/{id}/confirm":{"post":{"tags":["Gimnasio"],"operationId":"confirmReservation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationResponse"}}}}}}},"/api/reservations/{id}/cancel":{"post":{"tags":["Gimnasio"],"operationId":"cancelReservation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationResponse"}}}}}}},"/api/platform/organizations":{"get":{"tags":["Plataforma"],"operationId":"findAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrganizationResponse"}}}}}}},"post":{"tags":["Plataforma"],"summary":"Crear cliente (gimnasio)","description":"Registra la organización y crea un GYM_OWNER con ownerEmail/ownerPassword. Ese administrador inicia sesión en /api/auth/login y gestiona solo su gimnasio.","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrganizationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrganizationResponse"}}}}}}},"/api/packages":{"get":{"tags":["Gimnasio"],"operationId":"getPackages","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MembershipPackageResponse"}}}}}}},"post":{"tags":["Gimnasio"],"operationId":"createPackage","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MembershipPackageRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MembershipPackageResponse"}}}}}}},"/api/auth/register/{organizationId}":{"post":{"tags":["Autenticación"],"operationId":"register","parameters":[{"name":"organizationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["Autenticación"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/activities":{"get":{"tags":["Gimnasio"],"operationId":"getActivities","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ActivityResponse"}}}}}}},"post":{"tags":["Gimnasio"],"operationId":"createActivity","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActivityRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ActivityResponse"}}}}}}},"/api/activities/{activityId}/reservations":{"get":{"tags":["Gimnasio"],"operationId":"activityReservations","parameters":[{"name":"activityId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReservationResponse"}}}}}}},"post":{"tags":["Gimnasio"],"operationId":"createReservation","parameters":[{"name":"activityId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationResponse"}}}}}}},"/api/users/me":{"get":{"tags":["Gimnasio"],"operationId":"getMyProfile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/routines/me":{"get":{"tags":["Gimnasio"],"operationId":"myRoutines","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoutineResponse"}}}}}}}},"/api/reservations/me":{"get":{"tags":["Gimnasio"],"operationId":"myReservations","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReservationResponse"}}}}}}}},"/api/public/organizations":{"get":{"tags":["Público"],"operationId":"listActiveOrganizations","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrganizationResponse"}}}}}}}},"/api/activities/{id}":{"get":{"tags":["Gimnasio"],"operationId":"getActivity","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ActivityResponse"}}}}}}}},"components":{"schemas":{"UserCreateRequest":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string","description":"Opcional. Si se omite o está vacío, se usa 12345678"},"roles":{"type":"array","description":"Roles del usuario: GYM_OWNER (Admin), INSTRUCTOR, MEMBER","items":{"type":"string","enum":["PLATFORM_OWNER","GYM_OWNER","INSTRUCTOR","MEMBER"]}},"birthYear":{"type":"integer","format":"int32"},"age":{"type":"integer","format":"int32"},"goals":{"type":"string"},"phone":{"type":"string"}},"required":["email","firstName","lastName","roles"]},"MemberProfileResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"birthYear":{"type":"integer","format":"int32"},"age":{"type":"integer","format":"int32"},"goals":{"type":"string"},"phone":{"type":"string"},"emergencyContact":{"type":"string"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}},"organizationId":{"type":"integer","format":"int64"},"active":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"},"profile":{"$ref":"#/components/schemas/MemberProfileResponse"}}},"MemberProfileUpdateRequest":{"type":"object","properties":{"birthYear":{"type":"integer","format":"int32"},"age":{"type":"integer","format":"int32"},"goals":{"type":"string"},"phone":{"type":"string"},"emergencyContact":{"type":"string"}}},"RoutineRequestStatusUpdate":{"type":"object","properties":{"status":{"type":"string","enum":["PENDING","IN_PROGRESS","COMPLETED","REJECTED"]}},"required":["status"]},"RoutineRequestResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"memberId":{"type":"integer","format":"int64"},"memberName":{"type":"string"},"description":{"type":"string"},"goals":{"type":"string"},"status":{"type":"string"},"assignedInstructorId":{"type":"integer","format":"int64"},"assignedInstructorName":{"type":"string"}}},"OrganizationRequest":{"type":"object","properties":{"name":{"type":"string"},"slug":{"type":"string"},"ownerFirstName":{"type":"string"},"ownerLastName":{"type":"string"},"ownerEmail":{"type":"string"},"ownerPassword":{"type":"string","description":"Contraseña de acceso del administrador. Si se omite, se usa 12345678"},"contactEmail":{"type":"string"},"contactPhone":{"type":"string"},"subscriptionStatus":{"type":"string","enum":["ACTIVE","INACTIVE","TRIAL","SUSPENDED"]}},"required":["name","ownerEmail","ownerFirstName","ownerLastName","slug"]},"OrganizationResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"slug":{"type":"string"},"contactEmail":{"type":"string"},"contactPhone":{"type":"string"},"subscriptionStatus":{"type":"string","enum":["ACTIVE","INACTIVE","TRIAL","SUSPENDED"]},"active":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"},"ownerFirstName":{"type":"string"},"ownerLastName":{"type":"string"},"ownerEmail":{"type":"string"}}},"MembershipPackageRequest":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"durationMonths":{"type":"integer","format":"int32"},"addons":{"type":"array","items":{"$ref":"#/components/schemas/PackageAddonRequest"}}},"required":["name","price"]},"PackageAddonRequest":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"}},"required":["name","price"]},"MembershipPackageResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"durationMonths":{"type":"integer","format":"int32"},"active":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"},"addons":{"type":"array","items":{"$ref":"#/components/schemas/PackageAddonResponse"}}}},"PackageAddonResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"active":{"type":"boolean"}}},"CreateRoutineRequest":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"notes":{"type":"string"},"memberId":{"type":"integer","format":"int64"},"templateId":{"type":"integer","format":"int64"},"temporary":{"type":"boolean"},"exercises":{"type":"array","items":{"$ref":"#/components/schemas/RoutineExerciseRequest"}}},"required":["name"]},"RoutineExerciseRequest":{"type":"object","properties":{"exerciseName":{"type":"string"},"sets":{"type":"integer","format":"int32"},"reps":{"type":"integer","format":"int32"},"weight":{"type":"string"},"durationSeconds":{"type":"integer","format":"int32"},"notes":{"type":"string"},"orderIndex":{"type":"integer","format":"int32"}},"required":["exerciseName"]},"RoutineExerciseResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"exerciseName":{"type":"string"},"sets":{"type":"integer","format":"int32"},"reps":{"type":"integer","format":"int32"},"weight":{"type":"string"},"durationSeconds":{"type":"integer","format":"int32"},"notes":{"type":"string"},"orderIndex":{"type":"integer","format":"int32"}}},"RoutineResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"notes":{"type":"string"},"memberId":{"type":"integer","format":"int64"},"memberName":{"type":"string"},"instructorId":{"type":"integer","format":"int64"},"instructorName":{"type":"string"},"templateId":{"type":"integer","format":"int64"},"temporary":{"type":"boolean"},"exercises":{"type":"array","items":{"$ref":"#/components/schemas/RoutineExerciseResponse"}}}},"AssignTemplateRequest":{"type":"object","properties":{"templateId":{"type":"integer","format":"int64"},"memberIds":{"type":"array","items":{"type":"integer","format":"int64"}}}},"RoutineTemplateRequest":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"goal":{"type":"string"},"exercises":{"type":"array","items":{"$ref":"#/components/schemas/RoutineExerciseRequest"}}},"required":["name"]},"RoutineTemplateResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"goal":{"type":"string"},"instructorId":{"type":"integer","format":"int64"},"exercises":{"type":"array","items":{"$ref":"#/components/schemas/RoutineExerciseResponse"}}}},"RoutineRequestCreate":{"type":"object","properties":{"description":{"type":"string"},"goals":{"type":"string"}}},"ReservationResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"activityId":{"type":"integer","format":"int64"},"activityName":{"type":"string"},"memberId":{"type":"integer","format":"int64"},"memberName":{"type":"string"},"status":{"type":"string","enum":["PENDING","CONFIRMED","CANCELLED"]},"createdAt":{"type":"string","format":"date-time"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"},"userId":{"type":"integer","format":"int64"},"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}},"organizationId":{"type":"integer","format":"int64"}}},"ActivityRequest":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"activityDate":{"type":"string","format":"date"},"startTime":{"$ref":"#/components/schemas/LocalTime"},"endTime":{"$ref":"#/components/schemas/LocalTime"},"locationName":{"type":"string"},"instructorId":{"type":"integer","format":"int64"},"capacity":{"type":"integer","format":"int32"}},"required":["activityDate","name","startTime"]},"LocalTime":{"type":"object","properties":{"hour":{"type":"integer","format":"int32"},"minute":{"type":"integer","format":"int32"},"second":{"type":"integer","format":"int32"},"nano":{"type":"integer","format":"int32"}}},"ActivityResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"activityDate":{"type":"string","format":"date"},"startTime":{"$ref":"#/components/schemas/LocalTime"},"endTime":{"$ref":"#/components/schemas/LocalTime"},"locationName":{"type":"string"},"instructorId":{"type":"integer","format":"int64"},"instructorName":{"type":"string"},"capacity":{"type":"integer","format":"int32"},"confirmedReservations":{"type":"integer","format":"int64"},"hasCapacity":{"type":"boolean"},"active":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"}}}},"securitySchemes":{"bearerAuth":{"type":"http","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}